fix: running unit tests
