+++ src/app/app.component.spec.ts
@@ -1,30 +1,9 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { Component } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { MatAutocomplete } from '@angular/material/autocomplete';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AppComponent } from './app.component';
 import { ToObservablePipe } from './pipes';
-
-@Component({
-	selector: 'mat-label',
-})
-class MockMatLabelComponent {}
-
-@Component({
-	selector: 'mat-hint',
-})
-class MockMatHintComponent {}
-
-@Component({
-	selector: 'mat-error',
-})
-class MockMatErrorComponent {}
-
-@Component({
-	selector: 'mat-form-field',
-})
-class MockMatFormFieldComponent {}
 
 @Component({
 	selector: 'app-package-list',
@@ -40,17 +19,7 @@ describe('AppComponent', () => {
 	beforeEach(async () => {
 		await TestBed.configureTestingModule({
 			imports: [HttpClientTestingModule],
-			declarations: [
-				AppComponent,
-				ToObservablePipe,
-				MatAutocomplete, // Required for #auto="matAutocomplete" directive
-				MockMatLabelComponent,
-				MockMatHintComponent,
-				MockMatErrorComponent,
-				MockMatFormFieldComponent,
-				MockPackageListComponent,
-				MockDateRangePickerComponent,
-			],
+			declarations: [AppComponent, ToObservablePipe, MockPackageListComponent, MockDateRangePickerComponent],
 			providers: [
 				{
 					provide: MatSnackBar,
